requirements.txt
fastapi
uvicorn
onnxruntime
numpy
opencv-python-headless
Pillow
supervision
inference_sdk
python-multipart